Provide user access to configuration file path and improve configuration structure (#4377)
* move and add logging configuration logic to the configuration module

* update __init__ file

* update toml configuration file

* update yaml configuration file

* format

* Update custom implementation

* add __init__ docstring

* review suggestions

* move configure_logging to internal and add config_path

* format

* update changelog

* Update doc/releases/changelog-dev.md
