+++ pennylane/logging/log_config.yaml
@@ -19,7 +19,7 @@ handlers:
     class: logging.handlers.RotatingFileHandler
     level: INFO
     formatter: standard
-    filename: /tmp/logs/info.log
+    filename: tmp/logs/info.log
     maxBytes: 10485760 # 10MB
     backupCount: 20
     encoding: utf8
@@ -28,7 +28,7 @@ handlers:
     class: logging.handlers.RotatingFileHandler
     level: WARN
     formatter: standard
-    filename: /tmp/logs/warn.log
+    filename: tmp/logs/warn.log
     maxBytes: 10485760 # 10MB
     backupCount: 20
     encoding: utf8
@@ -37,7 +37,7 @@ handlers:
     class: logging.handlers.RotatingFileHandler
     level: ERROR
     formatter: detailed
-    filename: /tmp/logs/errors.log
+    filename: tmp/logs/errors.log
     maxBytes: 10485760 # 10MB
     backupCount: 20
     encoding: utf8
@@ -46,7 +46,7 @@ handlers:
     class: logging.handlers.RotatingFileHandler
     level: DEBUG
     formatter: standard
-    filename: /tmp/logs/debug.log
+    filename: tmp/logs/debug.log
     maxBytes: 10485760 # 10MB
     backupCount: 20
     encoding: utf8
@@ -55,7 +55,7 @@ handlers:
     class: logging.handlers.RotatingFileHandler
     level: DEBUG
     formatter: standard
-    filename: /tmp/logs/logs.log
+    filename: tmp/logs/logs.log
     maxBytes: 10485760 # 10MB
     backupCount: 20
     encoding: utf8
